Move old obc-classic tests to batch, removing redundancy
